@@ -1,6 +1,6 @@
 {
   "name": "vue-ts-decorate",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "A set of TypeScript decorators for make the work with Vue.js easiest.",
   "main": "./libs/index.js",
   "repository": {
@@ -8,9 +8,9 @@
     "url": "git://github.com/InDIOS/vue-ts-decorate.git"
   },
   "peerDependencies": {
-    "vue": "1.0.28 - 2.1.10"
+    "vue": "1.0.28 - 2.2.6"
   },
-  "typings": "./index",
+  "typings": "types/index.d.ts",
   "scripts": {
     "compile": "tsc -p .",
     "test:start": "karma start",
@@ -18,9 +18,7 @@
     "test": "npm run compile && npm run test:start"
   },
   "dependencies": {
-    "absurd-css": "^0.1.2",
-    "hash-sum": "^1.0.2",
-    "himalaya": "^0.2.2"
+    "himalaya": "*"
   },
   "author": "InDIOS",
   "license": "MIT",
@@ -33,8 +31,8 @@
     "karma-firefox-launcher": "^1.0.0",
     "karma-jasmine": "^1.0.2",
     "karma-spec-reporter": "0.0.26",
-    "typescript": "^2.0.6",
-    "vue": "1.0.28 - 2.1.10",
+    "typescript": "^2.3.2",
+    "vue": "^2.2.6",
     "watchify": "^3.7.0"
   }
 }
